debug

+++ src/com/company/ui/FullLength.java
@@ -156,7 +156,7 @@ public class FullLength extends JFrame {
             String answerFromText=answer.getText();
             //insert values
             QuestionDao questionDao=new QuestionDaoImp();
-            questionDao.create(num,SetAnExam.testNumber,comOrOp,"long",scoreBigDecimal,questionFromText,answerFromText);
+            questionDao.create((short)num,SetAnExam.testNumber,comOrOp,"long",scoreBigDecimal,questionFromText,answerFromText);
 
 
         });
@@ -174,7 +174,7 @@ public class FullLength extends JFrame {
             String answerFromText=answer.getText();
             //insert values
             QuestionDao questionDao=new QuestionDaoImp();
-            questionDao.create(num,SetAnExam.testNumber,comOrOp,"long",scoreBigDecimal,questionFromText,answerFromText);
+            questionDao.create((short)num,SetAnExam.testNumber,comOrOp,"long",scoreBigDecimal,questionFromText,answerFromText);
 
 
         });
